Pull quivr types to their own file, precovery_db.py was getting big
